% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{CenterOnMeanOfMean}
\alias{CenterOnMeanOfMean}
\title{CenterOnMeanOfMean}
\usage{
CenterOnMeanOfMean(d, centeredVar, idVariable = "id",
  groupingVariable = NULL)
}
\description{
center the variable for each subject then add the mean of mean
}
