% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{checkForPartialCorrelation}
\alias{checkForPartialCorrelation}
\title{checkForPartialCorrelation}
\usage{
checkForPartialCorrelation(data, variableCouplesDT, controledVariables,
  glmFamily = "gaussian", interaction = F, type = "glm", printOut = F,
  printFullGLM = T, normalize = F)
}
\description{
checkForPartialCorrelation
}
\note{
DevStatus: one pass - utility 4/5.
TODO(Albert): make example - discuss the methodology

Uses GLM to check the correlation of couples of variables store in a DT containing two columns v1 and v2
controlling for other variables specified in controled variables.

Normality assumption for variables : http://www.theanalysisfactor.com/dependent-variables-never-meet-normality/
}

