% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{circleFun}
\alias{circleFun}
\title{Circle drawing function}
\usage{
circleFun(center = c(0, 0), diameter = 1, npoints = 100)
}
\description{
http://stackoverflow.com/questions/6862742/draw-a-circle-with-ggplot2
}
