% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{compareToFullModel}
\alias{compareToFullModel}
\title{compareToFullModel}
\usage{
compareToFullModel(data, correlatedVariable, regressOverVariables,
  outputType = "store")
}
\arguments{
\item{data}{dta table containing the columns of data present in coreletedVariable and resgressedVariable}

\item{correlatedVariable}{: string vector with 2 rows contianing the correlated variable}

\item{regressOverVariables}{: string vector containing all variable names to regress on}
}
\description{
Performs a likelyhood ratio test lrtest on a simple model of correlatedVariable[1] ~ correlatedVariable[2] vs a full model
coreletedVariable[1] ~ correlatedVariable[2] + regressOverVariables...
}
\note{
: actually note sure if this function works
TODO(Albert): to test - change name - integrate to larger function
}
