% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{constraintVariable}
\alias{constraintVariable}
\title{constraintVariable}
\usage{
constraintVariable(data, idVariable, constrainedVariable, constrainingVariables,
  asManyObservations = 5, equalTo = NULL, differentFrom = NULL,
  replaceBy = NA)
}
\description{
Function that replace values of a certain constrained variables to a selected value (default to NA) for subject who did not fill certain constraints.
Constraints are described in term of a list of minimum number of observation either equal or differnt from a certain value. It is an OR testing.
}
