% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{correctForVariables}
\alias{correctForVariables}
\title{correctForVariables}
\usage{
correctForVariables(data, variables, correctVariables,
  regressOnAllVariables = F, normalizeVariables = T)
}
\description{
Correct each variables from the variables vector using the correctVariables vector for each row of data.
If regressOnAllVariables is T the coeficient are extraced from a regression with all variables in variables as well as correctVariables
- not recommended since colinear variable might produce strange results. Univariate regression by default.
Works only for continuous dependant variables.
}
\note{
DevStatus: one pass - utility 4/5
TODO(*): to discuss what is the best way to correct for variables \
}
