% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{crossValidatedPredictor}
\alias{crossValidatedPredictor}
\title{crossValidatedPredictor}
\usage{
crossValidatedPredictor(data, idVarname = NULL, dependantVariable,
  independantVariables, type = "leaveOneOut", family = "gaussian",
  kTimes = NULL, loopOverModels = T, preservedData = F, plotInfo = T,
  FTestWithRandomError = T)
}
\arguments{
\item{type}{= "leaveOneOut"}

\item{family}{= "gaussian"}

\item{kTimes}{= NULL}

\item{loopOverModels}{= T}

\item{preservedData}{= F}

\item{plotInfo}{= T}

\item{dVarname}{= NULL}

\item{TestWithRandomError}{= T}
}
\description{
Model selection tool with K-time cross validation and final validation on preserved data.
}
\note{
DevStatus: one pass - utility ?/5
TODO(Albert): Discuss methodology ++
}

