% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DensityPlot}
\alias{DensityPlot}
\title{DensityPlot}
\usage{
DensityPlot(d, dependantVar = NULL, groupingVar = NULL)
}
\description{
Plots the density of a dependantVar, with the option to group the plot on one groupingVar
}

