% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{eulersMethod}
\alias{eulersMethod}
\title{Euler's method for ODE first order}
\usage{
eulersMethod(a = 1, b = -1, c = 1, d = 0, x0 = 1, fx0 = 2,
  step = 0.5, goal = 3)
}
\description{
Euler's method for ODE first order
}
