% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionalityReduction.R
\name{faReady}
\alias{faReady}
\title{faReady}
\usage{
faReady(data, variables, numberOfComponents = NULL, toNumeric = T,
  looseSubjectsWithMissingValuesOverAtLeast = NULL)
}
\description{
After scaling the quantitative variables of the data table,
the function filters out subject with more than looseSubjectsWithMissingValuesOverAtLeast NAs.
It return the results of the fa function on those subjects for the specified variables.
}
\note{
Dev status : no pass yet - Usefulness: ?/5
}
