% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{fill_at}
\alias{fill_at}
\title{Fills a vector with either previous value if no fill is given or with the fill argument
if the predicate is true for the vector entry}
\usage{
fill_at(.v, .predicate = 0, .default = NA, .fill = NULL)
}
\arguments{
\item{.v}{vector to fill}

\item{.predicate}{either
an atomic in this case values that are equal to it will be filled
a formula: will be transformed to a function by calling as_mapper
a function: will be used as is}

\item{.default}{default value if no previous correct value was encountered}

\item{.fill}{if set all element of the vector for which the predicate is true will be replaced by
this value. If NULL, the previous correct value (or default if non encountered) will be used.}
}
\description{
Fills a vector with either previous value if no fill is given or with the fill argument
if the predicate is true for the vector entry
}
