% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{findExtremumsInRange}
\alias{findExtremumsInRange}
\title{getInterpolationFunction}
\usage{
findExtremumsInRange(splineFunc, range, delta, errorTolerance = 0.1,
  excapeDistance = NULL)
}
\description{
getInterpolationFunction
}
