% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatter.R
\docType{class}
\name{Formatter}
\alias{Formatter}
\title{Formatter Object}
\format{\code{\link{R6Class}} object.}
\usage{
Formatter
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class allowing to format data following a certain data_definition. If no data_definition is given, the varibale names
can be reformated to snake_case automatically using rename
}
\section{Methods}{

\describe{
  \item{\code{new(data = NULL, data_definition = NULL, rename = T, check_type = T)}}{This method is used to create an instance of the Formatter class }
  \item{\code{set_data (data, data_definition = NULL, auto_rename = F, check_type = T) }}{This method is used to set the data to be formatted.  }
  \item{\code{set_data_definition (data_definition = NULL, check_type = T)}}{This methods allows to setup the data definition propert and reformat the data accordingly. }
  \item{\code{get_data_definition ()}}{Returns current data_definition, if none is set it returns a default object. }
  }
}

\section{Fields}{

\describe{
     \item{\code{data_definition}}{data.table or data.frame defining the format of data.
Fields are : variable_name, formatted_name, variable_type, is_subject_id, is_group_variable, source (raw measurement : "raw" , or computed from: "variable_a,variable_b")
description: range values - how was it obtained - what does it represent - how can it be reproduced }
  }
}

\examples{
\dontrun{
data <- data.table(a = c(10,7), b = c("A", "B"))
data_definition <- data.table(variable_name = c("a", "b", "c"), formatted_name = c("a", "beta", "ceta"), variable_type = c("numeric", "factor", "character"), source = c("raw", "a", "raw"), description = c("0 to 10 - measure of awesomness", "A,B, or C - category of awesomnes computed from a", "just C"))
}

}
\keyword{data}
\keyword{format,}
\keyword{formatting,}
\keyword{naming,}
