% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{generate_log_interpolation_function}
\alias{generate_log_interpolation_function}
\title{Return a function interpolating point between (from_x, from_y) and (to_x, to_y)
following a logarithmic decay}
\usage{
generate_log_interpolation_function(from_x = 0, to_x = 0.05, from_y = 1,
  to_y = 0, t = 10)
}
\arguments{
\item{from_x}{numeric start point x}

\item{to_x}{numeric end point x}

\item{from_y}{numeric start point y}

\item{to_y}{numeric end point y}

\item{t}{numeric logarithmic slope}
}
\description{
Return a function interpolating point between (from_x, from_y) and (to_x, to_y)
following a logarithmic decay
}
