% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{getDensityForVector}
\alias{getDensityForVector}
\title{getDensityForVector}
\usage{
getDensityForVector(vector, variableName = "variable", ...)
}
\description{
Wrapper function for getDensityForX
PDF or CDF of a variable.

\code{atecFuncGetDensityForVector} is an helper function to call \code{getDensityForVector} with a vector.
}
