% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{getDensityForX}
\alias{getDensityForX}
\title{PDF or CDF of a variable.}
\usage{
getDensityForX(data, densityForVariableName, type = "PDF", nBins = NULL,
  meanOnVariable = NULL, globalInterval = NULL, zoomInterval = NULL,
  optionalVariableName = NULL, optionalVariableFormat = "density")
}
\arguments{
\item{optionalVariableFormat}{}
}
\value{
A data table containing two variable :
\describe{
     \item{varX}{contains the bins position on the variable axis}
     \item{histForX}{contains the bins density}
}
}
\description{
PDF or CDF of a variable.
}
\note{
DevStatus: One pass - utility: ?/5
TODO(Albert): Check if that function is not redundant - if not make it more robust 

\code{atecFuncGetDensityForX} returns a data.table containing either the PDF
or CDF of a variable.

Function takes a data table and return a data table containing Either a CDF
or a PDF for a variable. The data.table contains two variable varX the bin
position and histForX the bin probability. The variable of interest can be
average accross a grouping variable declared in meanOnVariable. Permits to
obtain precise distribution and interaction with an optional variable.

Zoom interval has not been added yet
If nBins is NULL number of bin = Number of observations/10
}

