% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{getDensityForXByGroupOnY}
\alias{getDensityForXByGroupOnY}
\title{getDensityForXByGroupOnY}
\usage{
getDensityForXByGroupOnY(data, densityForVariableName, groupByVariable,
  globalInterval = NULL, useGlobalInterval = FALSE, type = "PDF",
  nBins = NULL)
}
\value{
data.table with each distribution
}
\description{
Get the density distribution of a variable for several groups
}
\note{
DevStatus: one pass - utility 5/5
TODO(Albert): usability - name
}

