% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getDistanceFromLastOccurenceInVector}
\alias{getDistanceFromLastOccurenceInVector}
\title{getDistanceFromLastOccurenceInVector}
\usage{
getDistanceFromLastOccurenceInVector(vector, value,
  searchDirection = "before")
}
\description{
Returns a distance vector of the same size as input each eleemtn representing the distance from the last occurence of the value in the vector
}
