% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{GetSignificantPlotsXAgainstY}
\alias{GetSignificantPlotsXAgainstY}
\title{GetSignificantPlotsXAgainstY}
\usage{
GetSignificantPlotsXAgainstY(aovTest, variableNames, plotPrefix = NULL)
}
\description{
Extract information from the Anova contained in the output of PlotXAgainstY to produce a vector of plot names better fit to show those significant data
}
