% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionalityReduction.R
\name{PCAReady}
\alias{PCAReady}
\title{PCAReady}
\usage{
PCAReady(data, variables, numberOfComponents = NULL, principalOnly = T,
  toNumeric = T, rotation = T,
  looseSubjectsWithMissingValuesOverAtLeast = NULL)
}
\description{
does fa analysis if numberOfComponents is NULL. Rotate varimax default.
}
\note{
Dev status: one pass - to clean -- Usefulness: 4/5

Produce NA safe PCA using principal, as well as PCA from factominr.
}
