% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_correlation_ci}
\alias{plot_correlation_ci}
\title{Does the correlations between left and right vars using corr.test
Then plots a vertical graph with correlations confidence intervals.}
\usage{
plot_correlation_ci(data, left_vars, right_vars, .adjust = "holm",
  .method = "spearman", .xlab = "Correlations CI",
  .ylab = "Left variables", .labels_left = NULL,
  .right_var_legend_name = NULL, ...)
}
\arguments{
\item{data}{data.frame contains all the columns to correlate}

\item{left_vars}{character contains the variable names to plot on the left}

\item{right_vars}{character contains the variable names to plot on the right (nested)}

\item{.adjust}{character adjust procedure default to holm (see corr.test)}

\item{.method}{character statistical method default to spearman}

\item{.xlab}{character x label name}

\item{.ylab}{character y label name (left variables)}

\item{.right_var_legend_name}{character name for the right_variable legend}
}
\description{
Does the correlations between left and right vars using corr.test
Then plots a vertical graph with correlations confidence intervals.
}
