% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{PlotEachCorrelation}
\alias{PlotEachCorrelation}
\title{PlotEachCorrelation(data, couples)}
\usage{
PlotEachCorrelation(data, couples, forPValueInferiorTo = "0.05",
  noPrint = F)
}
\arguments{
\item{data}{a data.table}

\item{couples}{data.table needs to contain 2 columns named var1 and var2}

\item{forPValueInferiorTo}{numeric value for the significance threshold}

\item{noPrint}{boolean. If set to false no plots a build and only text and a final a correlation table is produce}
}
\description{
Exploratory data analysis function testing correlation between variable couples, using adapted statistical methodology based on the variable types.
Note: variable are categorized mainly by the isContinuous function - discrete variable are treated as categorical
}
\note{
DevStatus: second pass - utility 5/5
TODO(Albert): Discuss methodology ++, Discrete as categorical (as underlying phenomena could drive discrete...)
}
