% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotMeanSliding}
\alias{PlotMeanSliding}
\title{PlotMeanSliding}
\usage{
PlotMeanSliding(data, variable, slideOverVariable, groupBy = NULL,
  printTable = F, nbins = 10, range = NULL)
}
\description{
Plot the mean with error bars
}

