% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotOutlier}
\alias{PlotOutlier}
\title{PlotOutlier}
\usage{
PlotOutlier(data, idValue, idVariable = "id", variablesToPlot = NULL)
}
\arguments{
\item{idValue}{Either a numeric value indicating the row, or a string that will be
match to the idVariable}

\item{idVariable}{Optional, A string containing the name of the idValue variable
used for string matching with idValue. Default with "id"}
}
\description{
Outlier visualization.
}
\details{
\code{PlotOutlier} plots a specified outlier on various variable in
the dataset.
}

