% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotXAgainstY}
\alias{PlotXAgainstY}
\title{PlotXAgainstY}
\usage{
PlotXAgainstY(dependantVar, groupingVars, data, outputType = "store",
  addNameOfDependantVarToPlot = F)
}
\description{
If outputType == "store" (default)
--- Returns in [[1]] ANOVA test for DependantVar ~ groupingVar[1]*groupingVar[2]*...
--- Returns in [[2]] various plots with descriptive name
--- Returns in [[3]] the names of plots in [[2]] that are significant
--- Returns in [[4]] the names of the dependant variables that were significant
If outputType == "print" the function printsall anova summaries and all plots
If outputType == "printSignificant" the functions only print significant results
}
\details{
Produce an ANOVA and graphs for dependantVar ~ groupingVars, in data
}

