% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{rollsmooth}
\alias{rollsmooth}
\title{Applies a function using a bilateral window step = 1}
\usage{
rollsmooth(v, width, funs = mean, ...)
}
\arguments{
\item{v}{vector upon which to apply funs}

\item{width}{size of the window}

\item{funs}{function to apply to the vector}

\item{...}{additional parameters to pass to funs}
}
\description{
Applies a function using a bilateral window step = 1
}
