% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{scrubNA}
\alias{scrubNA}
\title{scrubNA}
\usage{
scrubNA(data, variableNames = NULL, duplicateData = T)
}
\arguments{
\item{data}{Contains a data.table or a data.frame.}

\item{variableNames}{Vector containing the names of the variables to scrub.
If left empty, the algorithm will try to scrub the columns}
}
\description{
scrubNA
}
