% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{setNAToX}
\alias{setNAToX}
\title{atecFuncSetNAToXFromDT}
\usage{
setNAToX(DT, X = 0)
}
\description{
atecFuncSetNAToXFromDT
}
\note{
DevStatus: one pass - utility 5/5

Sets NA to a certain value X in a DT (all columns)
}
