% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid function}
\usage{
sigmoid(x, params = c(1, 1, 0), type = "exp")
}
\description{
credits: Kyriakos Chatzidimitriou
http://kyrcha.info/2012/07/08/tutorials-fitting-a-sigmoid-function-in-r/
}
