% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{slideVector}
\alias{slideVector}
\title{Sliding vector}
\usage{
slideVector(vector, by, fillIn = 0, keepSameLength = T)
}
\arguments{
\item{by}{numeric value indicating by how many indices the vector has to be slided}
}
\description{
Slides vector by a certain amount of indices.
}
