% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{SummaryToText}
\alias{SummaryToText}
\title{SummaryToText BNICE
TODO(Albert): Example - Put in BeNice? - make it more general able to summarize all kinds of tests}
\usage{
SummaryToText(summaryTable, anovaTest = NULL, type = "CI", precision = 2,
  printOut = F, significance = 0.05)
}
\arguments{
\item{summaryTable}{A data table following the classic summarySE format group N value sd se and ci}

\item{type}{Either CI or SD}
}
\description{
Produce text description of 95% CI APA style from a summary table, or an anovaTest.
}

