% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{timeSeriesTotalVariation}
\alias{timeSeriesTotalVariation}
\title{timeSeriesTotalVariation}
\usage{
timeSeriesTotalVariation(timeSeries)
}
\description{
timeSeriesTotalVariation
}
\note{
DevStatus: one pass - utility 3/5
TODO(Albert): Integrate in more general time series analysis tools

Function to calculate the total variation of a time series contained in a vector timeSeries
}
