% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{toNumeric}
\alias{toNumeric}
\title{toNumeric}
\usage{
toNumeric(data, variables = NULL)
}
\description{
Safely fransforms variable in a data table as.numeric.
}
\note{
DevStatus: one pass - utility 2/5
}
