% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{VariableTableDescription}
\alias{VariableTableDescription}
\title{VariableTableDescription
TODO(Albert): Put in BeNice ??}
\usage{
VariableTableDescription(data, variables, idVariable = "id", precision = 2)
}
\description{
Creates an HTML table to describe one or several variables in terms of Mean, SD, Range, Skewness, Kurtosis, Internal consistency methodology varies for variable types and will not be included.
Previously computed column can be added in the form of a vector of a matrix (m variables * n columns)
}

