% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vectorRatioOfSimilarity}
\alias{vectorRatioOfSimilarity}
\title{vectorRatioOfSimilarity}
\usage{
vectorRatioOfSimilarity(vector1, vector2, na.rm = T)
}
\description{
returns the point to point ratio of similarity between two vectors
}
\details{
@description if vector1 = vector2 - return 1. If no index  i exist such as vector1[i] == vector2[i] returns 0
}
