% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Winsorize}
\alias{Winsorize}
\title{Winsorize a vector}
\usage{
Winsorize(x, minval = NULL, maxval = NULL, probs = c(0.05, 0.95),
  na.rm = FALSE)
}
\description{
Permits to set a threshold for extreme values over a certain quantile. Outlier correction technique.
}
