% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{writeTemp}
\alias{writeTemp}
\title{writeTemp}
\usage{
writeTemp(data, fileName)
}
\description{
writeTemp
}
\note{
DevStatus: no pass - utility ?/5 - should probably be part of better organized functions keeping tracks of data versions and folder structures 

Write a data frame contained into an xls file in the temporary data folder
}
